+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet+PaymentMethodAvailability.swift
@@ -30,7 +30,7 @@ extension PaymentSheet {
         .grabPay,
         .FPX,
         .alipay,
-        .OXXO, .zip, .revolutPay, .amazonPay, .mobilePay, .paynow,
+        .OXXO, .zip, .revolutPay, .amazonPay, .mobilePay, .konbini, .paynow,
     ]
 
     /// An unordered list of paymentMethodtypes that can be used with Link in PaymentSheet