+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentSheetLPMConfirmFlowTests.swift
@@ -31,6 +31,7 @@ final class PaymentSheet_LPM_ConfirmFlowTests: XCTestCase {
         case BE = "be"
         case MX = "mex"  // The CI Backend uses "mex" instead of "mx"
         case AU = "au"
+        case JP = "jp"
 
         var publishableKey: String {
             switch self {
@@ -46,6 +47,8 @@ final class PaymentSheet_LPM_ConfirmFlowTests: XCTestCase {
                 return STPTestingMEXPublishableKey
             case .AU:
                 return STPTestingAUPublishableKey
+            case .JP:
+                return STPTestingJPPublishableKey
             }
         }
     }
@@ -184,6 +187,16 @@ final class PaymentSheet_LPM_ConfirmFlowTests: XCTestCase {
         }
     }
 
+    func testKonbiniConfirmFlows() async throws {
+        try await _testConfirm(intentKinds: [.paymentIntent],
+                               currency: "JPY",
+                               paymentMethodType: .dynamic("konbini"),
+                               merchantCountry: .JP) { form in
+            form.getTextFieldElement("Full name")?.setText("Jane Doe")
+            form.getTextFieldElement("Email")?.setText("foo@bar.com")
+        }
+    }
+
     func testPayNowConfirmFlows() async throws {
         try await _testConfirm(intentKinds: [.paymentIntent],
                                currency: "SGD",
@@ -191,8 +204,7 @@ final class PaymentSheet_LPM_ConfirmFlowTests: XCTestCase {
                                merchantCountry: .SG) { form in
             // PayNow has no input fields
             XCTAssertEqual(form.getAllSubElements().count, 1)
-        }
-    }
+    }        
 }
 
 // MARK: - Helper methods